move from test_omegaconf.py to test_utils.py
